# Optimize inside_points with Numpy magic