--- a/test_icosphere.py
+++ b/test_icosphere.py
@@ -1,0 +1,11 @@
+import pytest
+
+import icosphere
+
+
+@pytest.mark.parametrize("nu", [1, 2, 3, 5, 10])
+def test_icosphere(nu):
+    sv, sf = icosphere.icosphere(nu)
+    # Just a smoke test for now...
+    assert sv.shape
+    assert sf.shape