I profiled the library a bit and noticed `inside_points()` was by far the hottest function.

I wrote a small verification/benchmark script that tests that the output for `icosphere(nu)` (with `nu` from 5 to 200) is this is the same as on be65c259ed09749c02b1d625d774f7b416480205.

The approximate speedup on my machine is 5-8x (and I'm not a Numpy magician by any means, so there could be ways to make this faster still):

```
nu = 5 / ratio: 0.97 (this 0.01s / baseline 0.01s)
nu = 10 / ratio: 0.71 (this 0.01s / baseline 0.02s)
nu = 15 / ratio: 0.55 (this 0.02s / baseline 0.03s)
nu = 20 / ratio: 0.39 (this 0.02s / baseline 0.05s)
nu = 25 / ratio: 0.49 (this 0.04s / baseline 0.07s)
nu = 30 / ratio: 0.34 (this 0.03s / baseline 0.10s)
nu = 35 / ratio: 0.29 (this 0.04s / baseline 0.13s)
nu = 40 / ratio: 0.24 (this 0.04s / baseline 0.18s)
nu = 45 / ratio: 0.22 (this 0.05s / baseline 0.22s)
nu = 50 / ratio: 0.20 (this 0.05s / baseline 0.27s)
nu = 55 / ratio: 0.20 (this 0.06s / baseline 0.30s)
nu = 60 / ratio: 0.19 (this 0.07s / baseline 0.36s)
nu = 65 / ratio: 0.17 (this 0.07s / baseline 0.43s)
nu = 70 / ratio: 0.16 (this 0.08s / baseline 0.50s)
[...]
nu = 155 / ratio: 0.13 (this 0.33s / baseline 2.55s)
nu = 160 / ratio: 0.10 (this 0.29s / baseline 2.88s)
nu = 165 / ratio: 0.11 (this 0.32s / baseline 2.94s)
nu = 170 / ratio: 0.12 (this 0.36s / baseline 3.08s)
nu = 175 / ratio: 0.11 (this 0.34s / baseline 3.18s)
nu = 180 / ratio: 0.12 (this 0.41s / baseline 3.36s)
nu = 185 / ratio: 0.11 (this 0.39s / baseline 3.67s)
nu = 190 / ratio: 0.10 (this 0.40s / baseline 3.85s)
nu = 195 / ratio: 0.11 (this 0.45s / baseline 4.07s)
nu = 200 / ratio: 0.11 (this 0.45s / baseline 4.30s)
nu = 205 / ratio: 0.10 (this 0.45s / baseline 4.67s)
nu = 210 / ratio: 0.10 (this 0.48s / baseline 5.01s)
nu = 215 / ratio: 0.11 (this 0.51s / baseline 4.82s)
```