@jacobbieker 2023-12-06
This seems to fail on `nu=2`, I get an error of ```ValueError: need at least one array to concatenate```