@akx 2023-12-07
@jacobbieker Good catch (and nice to see someone's trying this code). Pushed a commit to fix that (basically special-casing when `vAB` only has one vertex so far).