@jacobbieker 2023-12-07
Thanks for the update! This has been really useful for https://github.com/openclimatefix/graph_weather/issues/55 where we are reimplementing GraphCast for weather forecasting.